feat(cache-schema): add atomic schemas
Split standard schema implementations by Rust type family so scalars, tuples, NonZero wrappers, and atomics have separate ownership. Atomic integer and bool schemas are intentionally distinct from their inner scalar values.
