+++ crates/ssa-cache-schema/src/impls/mod.rs
@@ -1,0 +1,8 @@
+mod atomics;
+mod collections;
+mod nonzero;
+mod scalars;
+mod sequences;
+mod text;
+mod transparent;
+mod tuples;
